+++ python/lsst/ts/mtdome/mock_llc/lwscs.py
@@ -186,8 +186,7 @@ class LwscsStatus(BaseMockStatus):
             tests it can be convenient to use other values.
 
         """
-        self.elevation_motion.stop(start_tai)
-        self.duration = 0.0
+        self.duration = self.elevation_motion.stop(start_tai)
         return self.duration
 
     async def go_stationary(self, start_tai: float) -> float:
@@ -202,8 +201,7 @@ class LwscsStatus(BaseMockStatus):
             tests it can be convenient to use other values.
 
         """
-        self.elevation_motion.go_stationary(start_tai)
-        self.duration = 0.0
+        self.duration = self.elevation_motion.go_stationary(start_tai)
         return self.duration
 
     async def exit_fault(self, start_tai: float) -> float: