+++ python/lsst/ts/mtdome/mock_llc/mock_motion/elevation_motion.py
@@ -26,7 +26,7 @@ import math
 import typing
 
 from .base_llc_motion import BaseLlcMotion
-from ...enums import IntermediateState, LlcMotionState
+from ...enums import LlcMotionState
 from lsst.ts import utils
 
 
@@ -101,17 +101,13 @@ class ElevationMotion(BaseLlcMotion):
         """
         if tai >= self._end_tai:
             if self._commanded_motion_state in [
-                LlcMotionState.STOPPING,
                 LlcMotionState.STOPPED,
                 LlcMotionState.MOVING,
             ]:
                 motion_state = LlcMotionState.STOPPED
                 position = self._end_position
                 velocity = 0.0
-            elif self._commanded_motion_state in [
-                IntermediateState.GO_STATIONARY,
-                LlcMotionState.STATIONARY,
-            ]:
+            elif self._commanded_motion_state == LlcMotionState.STATIONARY:
                 motion_state = LlcMotionState.STATIONARY
                 position = self._end_position
                 velocity = 0.0
@@ -132,15 +128,9 @@ class ElevationMotion(BaseLlcMotion):
                     motion_state = LlcMotionState.STOPPED
                     velocity = 0.0
                 if self._crawl_velocity == 0.0:
-                    if self._commanded_motion_state in [
-                        LlcMotionState.STOPPING,
-                        LlcMotionState.STOPPED,
-                    ]:
+                    if self._commanded_motion_state == LlcMotionState.STOPPED:
                         motion_state = LlcMotionState.STOPPED
-                    elif self._commanded_motion_state in [
-                        IntermediateState.GO_STATIONARY,
-                        LlcMotionState.STATIONARY,
-                    ]:
+                    elif self._commanded_motion_state == LlcMotionState.STATIONARY:
                         motion_state = LlcMotionState.STATIONARY
 
                     velocity = 0.0
@@ -155,10 +145,10 @@ class ElevationMotion(BaseLlcMotion):
             velocity = self._max_speed
             if distance < 0.0:
                 velocity = -self._max_speed
-            if self._commanded_motion_state == LlcMotionState.STOPPING:
+            if self._commanded_motion_state == LlcMotionState.STOPPED:
                 motion_state = LlcMotionState.STOPPED
                 velocity = 0.0
-            elif self._commanded_motion_state == IntermediateState.GO_STATIONARY:
+            elif self._commanded_motion_state == LlcMotionState.STATIONARY:
                 motion_state = LlcMotionState.STATIONARY
                 velocity = 0.0
             else:
@@ -167,7 +157,7 @@ class ElevationMotion(BaseLlcMotion):
         position = utils.angle_wrap_nonnegative(math.degrees(position)).rad
         return position, velocity, motion_state
 
-    def stop(self, start_tai: float) -> None:
+    def stop(self, start_tai: float) -> float:
         """Stops the current motion.
 
         Parameters
@@ -184,9 +174,10 @@ class ElevationMotion(BaseLlcMotion):
         self._start_position = position
         self._end_position = position
         self._crawl_velocity = 0
-        self._commanded_motion_state = LlcMotionState.STOPPING
+        self._commanded_motion_state = LlcMotionState.STOPPED
+        return 0.0
 
-    def go_stationary(self, start_tai: float) -> None:
+    def go_stationary(self, start_tai: float) -> float:
         """Go to stationary state.
 
         Parameters
@@ -203,7 +194,8 @@ class ElevationMotion(BaseLlcMotion):
         self._start_position = position
         self._end_position = position
         self._crawl_velocity = 0.0
-        self._commanded_motion_state = IntermediateState.GO_STATIONARY
+        self._commanded_motion_state = LlcMotionState.STATIONARY
+        return 0.0
 
     def park(self, start_tai: float) -> float:
         """Not used for the elevation motion.
@@ -239,4 +231,4 @@ class ElevationMotion(BaseLlcMotion):
         self._start_position = position
         self._end_position = position
         self._crawl_velocity = 0.0
-        self._commanded_motion_state = IntermediateState.GO_STATIONARY
+        self._commanded_motion_state = LlcMotionState.STATIONARY