+++ doc/version_history.rst
@@ -14,6 +14,7 @@ Changes:
 * Added "appliedConfiguration" to the status replies of AMCS and LWSCS.
 * Added additional LLC states that are translated into MOVING, STOPPING or PARKING.
 * Updated the documentation to contain the full list of LLC commands and statuses.
+* Updated the MockController to handle the new state transitions.
 
 Requires: